fix: Use static MIT license badge until LICENSE file is on main branch
The GitHub license detection badge requires the LICENSE file to be on the
main branch. Since our LICENSE file is currently only on the dev branch,
using a static MIT badge that will work immediately.

Once the LICENSE file is merged to main, we can switch back to the
automatic GitHub license detection badge if desired.